Moves ObjectType and Entity to two different classes
ObjectType and Entity were created on the basis on the one shared class.
The relations were introduced to the database and to describe them I need
to use some additional fields for Entity. These fields are not useful
in ObjectType and can cause some bugs. That's why two classes are introduced -
one for Entity and another one for ObjectType.